build.rs,localtime.rs: Hardcode libtz's `time_t` to 64 bits.
I _think_ this is the right solution. Since it doesn't affect system time_t, there
aren't any ABI issues to worry about. It means the whole library, as is, is
ready for 2038.

This ended up changing all the exported symbols in the library, prefixing them
with `tz_`. I removed them with the `link_name` attribute because the extra
`tz_` is ugly (and redundant since they're all prefixed by the crate anyway).