build.rs,Cargo.toml: Switch to the `cc` crate as it seems more robust
It knows how to set up the basic compiler flags across disparate
machines (-fPIC in particular), where the `make` crate just runs
`make` and prays for the best.

This ends up duplicating some basic parts of the Makefile and might
need updating if the Makefile ever adds a new file, but we'll get over
it.